refactor: Events index is function component
The redux app state has several branches, including the "events" branch.
Previously, this "events" branch was provided as props to the Events component through the the `connect` call and its `mapStateToProps` argument.
Another prop provided by the `connect` call was the `getAll()` prop.
This prop didn't rely on a branch of the app state; instead it was set up to dispatch actions which were then handled by the redux action handlers.

We still select the same "events" branch of the app state, but do so within the component itself through the `useSelector` call, which receives the entire app state, and slices-off and returns the "events" branch.

We still dispatch the same "getAll" action when the component mounts, but do so through the `useDispatch` call, which returns the redux dispatcher directly.

Instead of using the "componentDidMount" Component lifecycle hook, we use the `useEffect` hook to run the hook once, when the component is first rendered.

The second argument to "connect" was a mapping of action creators to prop names. Within that call, redux wrapped each of these action creators with the equivalent of `dispatch(…)`.
Now, we have access to `dispatch` directly, through the `useDispatch` hook.